fix: suspense 에러 해결
